requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq>=1.2.0
PyPDF2==3.0.1
python-docx==1.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
mysql-connector-python==8.2.0
